refactor(team): extract prompts.go + mcp_config.go (C13)
Mechanical file split. Two themed files extracted from launcher.go:

* prompts.go (94 lines, new) — buildPrompt, newPromptBuilder,
  writeAgentPromptFile, cleanupAgentTempFiles, resolvePermissionFlags.

* mcp_config.go (178 lines, new) — buildMCPServerMap (composes the
  MCP server entry), ensureMCPConfig (writes the team-wide config),
  ensureAgentMCPConfig (writes a per-agent filtered config),
  agentMCPServers, codingAgentSlugs.

claudeCommand stays in launcher.go because moving it would trigger
the no-secrets pre-commit hook on the ONE_SECRET= env-var assembly
(false positive: the regex matches `secret=` inside the literal).
A future PR can move it once the hook regex is loosened.

launcher.go: 1523 -> 1290 lines (-233). Cumulative since main:
4998 -> 1290 = -3708 lines (-74.2%).

Stacked on refactor/launcher-c12-lifecycle-split (C12).
