+++ internal/team/launcher.go
@@ -17,7 +17,6 @@ import (
 	"os"
 	"os/exec"
 	"path/filepath"
-	"sort"
 	"strings"
 	"sync"
 	"sync/atomic"
@@ -918,43 +917,6 @@ func (l *Launcher) recordPaneSpawnFailure(slug, reason string) {
 	l.failedPaneSlugs[slug] = reason
 }
 
-// buildPrompt generates the system prompt for an agent. The body lives on
-// promptBuilder (see prompt_builder.go) so it can be tested without a
-// Launcher; this wrapper assembles the snapshot accessors from launcher
-// state and delegates.
-func (l *Launcher) buildPrompt(slug string) string {
-	return l.newPromptBuilder().Build(slug)
-}
-
-// newPromptBuilder captures the launcher state the prompt depends on into
-// a promptBuilder. Built fresh per call so each prompt sees the current
-// member roster and active policies; the construction itself is cheap
-// (closures + a couple of map lookups).
-func (l *Launcher) newPromptBuilder() *promptBuilder {
-	memoryBackend := config.ResolveMemoryBackend("")
-	noNex := config.ResolveNoNex() || config.ResolveAPIKey("") == ""
-	return &promptBuilder{
-		isOneOnOne:  l.isOneOnOne,
-		isFocusMode: l.isFocusModeEnabled,
-		packName:    l.PackName,
-		leadSlug:    l.targeter().LeadSlug,
-		members:     l.officeMembersSnapshot,
-		policies: func() []officePolicy {
-			if l == nil || l.broker == nil {
-				return nil
-			}
-			policies := l.broker.ListPolicies()
-			// Sort so the policies section is deterministic and
-			// cache-friendly across turns.
-			sort.Slice(policies, func(i, j int) bool { return policies[i].ID < policies[j].ID })
-			return policies
-		},
-		nameFor:        l.targeter().NameFor,
-		markdownMemory: memoryBackend == config.MemoryBackendMarkdown,
-		nexDisabled:    noNex,
-	}
-}
-
 // claudeCommand builds the shell command string for spawning a claude session.
 // Sets WUPHF_AGENT_SLUG so the MCP knows which agent this session serves.
 // claudeCommand returns the shell command that launches an interactive
@@ -1027,201 +989,6 @@ func (l *Launcher) claudeCommand(slug, systemPrompt string) (string, error) {
 		mcpConfig,
 		name,
 	), nil
-}
-
-// writeAgentPromptFile persists the per-agent system prompt to a stable
-// per-slug temp file so it can be passed to `claude --append-system-prompt-file`
-// without bloating the tmux command.
-//
-// File naming mirrors ensureAgentMCPConfig (wuphf-mcp-<slug>.json) so both
-// artifacts are easy to clean up together. Perms are 0o600 because the prompt
-// can contain team-internal instructions and tool lists.
-func (l *Launcher) writeAgentPromptFile(slug, prompt string) (string, error) {
-	path := filepath.Join(os.TempDir(), "wuphf-prompt-"+slug+".txt")
-	if err := os.WriteFile(path, []byte(prompt), 0o600); err != nil {
-		return "", err
-	}
-	return path, nil
-}
-
-// cleanupAgentTempFiles removes the per-agent MCP config + system prompt
-// temp files for every known office member. Safe to call multiple times and
-// idempotent — missing files are ignored. Called from Shutdown so the broker
-// token + prompt content do not linger in $TMPDIR after the session ends.
-func (l *Launcher) cleanupAgentTempFiles() {
-	tmp := os.TempDir()
-	for _, m := range l.officeMembersSnapshot() {
-		for _, path := range []string{
-			filepath.Join(tmp, "wuphf-mcp-"+m.Slug+".json"),
-			filepath.Join(tmp, "wuphf-prompt-"+m.Slug+".txt"),
-		} {
-			_ = os.Remove(path)
-		}
-	}
-}
-
-// resolvePermissionFlags returns the Claude Code permission flags for an agent.
-// All agents run in bypass mode by default — the team is autonomous.
-func (l *Launcher) resolvePermissionFlags(slug string) string {
-	return "--permission-mode bypassPermissions --dangerously-skip-permissions"
-}
-
-// codingAgentSlugs lists agents that default to a minimal coding-focused MCP set.
-// Task-level local_worktree isolation is driven by execution_mode, not this list.
-var codingAgentSlugs = map[string]bool{
-	"eng":       true,
-	"fe":        true,
-	"be":        true,
-	"ai":        true,
-	"qa":        true,
-	"tech-lead": true,
-}
-
-// agentMCPServers returns the MCP server keys that a given agent should receive.
-func agentMCPServers(slug string) []string {
-	channel := strings.TrimSpace(os.Getenv("WUPHF_CHANNEL"))
-	// DM mode: only wuphf-office (minimal tool set, no nex overhead)
-	if strings.HasPrefix(channel, "dm-") {
-		return []string{"wuphf-office"}
-	}
-	if codingAgentSlugs[slug] {
-		return []string{"wuphf-office"}
-	}
-	return []string{"wuphf-office", "nex"}
-}
-
-// buildMCPServerMap constructs the full set of MCP server entries.
-// This is the shared helper used by both ensureMCPConfig and ensureAgentMCPConfig.
-func (l *Launcher) buildMCPServerMap() (map[string]any, error) {
-	apiKey := config.ResolveAPIKey("")
-	servers := map[string]any{}
-	wuphfBinary, err := os.Executable()
-	if err != nil {
-		return nil, err
-	}
-
-	office := map[string]any{
-		"command": wuphfBinary,
-		"args":    []string{"mcp-team"},
-	}
-	servers["wuphf-office"] = office
-	if oneSecret := strings.TrimSpace(config.ResolveOneSecret()); oneSecret != "" {
-		office["env"] = map[string]string{
-			"ONE_SECRET": oneSecret,
-		}
-	}
-	if identity := strings.TrimSpace(config.ResolveOneIdentity()); identity != "" {
-		env, _ := office["env"].(map[string]string)
-		if env == nil {
-			env = map[string]string{}
-		}
-		env["ONE_IDENTITY"] = identity
-		if identityType := strings.TrimSpace(config.ResolveOneIdentityType()); identityType != "" {
-			env["ONE_IDENTITY_TYPE"] = identityType
-		}
-		office["env"] = env
-	}
-
-	switch config.ResolveMemoryBackend("") {
-	case config.MemoryBackendNex:
-		if apiKey != "" {
-			env, _ := office["env"].(map[string]string)
-			if env == nil {
-				env = map[string]string{}
-			}
-			env["WUPHF_API_KEY"] = apiKey
-			env["NEX_API_KEY"] = apiKey
-			office["env"] = env
-		}
-	case config.MemoryBackendGBrain:
-		env, _ := office["env"].(map[string]string)
-		if env == nil {
-			env = map[string]string{}
-		}
-		for key, value := range gbrainMCPEnv() {
-			env[key] = value
-		}
-		office["env"] = env
-	}
-
-	if memoryServer, err := resolvedMemoryMCPServer(); err != nil {
-		return nil, err
-	} else if memoryServer != nil && len(memoryServer.Env) > 0 {
-		env, _ := office["env"].(map[string]string)
-		if env == nil {
-			env = map[string]string{}
-		}
-		for key, value := range memoryServer.Env {
-			env[key] = value
-		}
-		office["env"] = env
-	}
-
-	if !config.ResolveNoNex() && apiKey != "" {
-		if nexMCP, err := exec.LookPath("nex-mcp"); err == nil {
-			servers["nex"] = map[string]any{
-				"command": nexMCP,
-				"env": map[string]string{
-					"WUPHF_API_KEY": apiKey,
-					"NEX_API_KEY":   apiKey,
-				},
-			}
-		}
-	}
-
-	return servers, nil
-}
-
-func (l *Launcher) ensureMCPConfig() (string, error) {
-	servers, err := l.buildMCPServerMap()
-	if err != nil {
-		return "", err
-	}
-
-	cfg := map[string]any{
-		"mcpServers": servers,
-	}
-	data, err := json.Marshal(cfg)
-	if err != nil {
-		return "", err
-	}
-
-	path := filepath.Join(os.TempDir(), "wuphf-team-mcp.json")
-	if err := os.WriteFile(path, data, 0o600); err != nil {
-		return "", err
-	}
-	return path, nil
-}
-
-// ensureAgentMCPConfig writes a per-agent MCP config containing only the servers
-// that agent needs. Returns the config file path.
-func (l *Launcher) ensureAgentMCPConfig(slug string) (string, error) {
-	allServers, err := l.buildMCPServerMap()
-	if err != nil {
-		return "", err
-	}
-
-	allowed := agentMCPServers(slug)
-	filtered := make(map[string]any, len(allowed))
-	for _, key := range allowed {
-		if srv, ok := allServers[key]; ok {
-			filtered[key] = srv
-		}
-	}
-
-	cfg := map[string]any{
-		"mcpServers": filtered,
-	}
-	data, err := json.Marshal(cfg)
-	if err != nil {
-		return "", err
-	}
-
-	path := filepath.Join(os.TempDir(), "wuphf-mcp-"+slug+".json")
-	if err := os.WriteFile(path, data, 0o600); err != nil {
-		return "", err
-	}
-	return path, nil
 }
 
 // agentActiveTask returns the first in_progress task owned by the given agent slug.
